feat: 서비스로 로직 이동 후 DB 연동
